plant update

+++ app/Http/Controllers/Api/PlantController.php
@@ -1,9 +1,11 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
-use App\Models\Plant;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Models\Plant;
+use Illuminate\Support\Facades\Storage;
 
 class PlantController extends Controller
 {
@@ -11,14 +13,6 @@ class PlantController extends Controller
      * Display a listing of the resource.
      */
     public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
     {
         //
     }
@@ -34,15 +28,7 @@ class PlantController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Plant $plant)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Plant $plant)
+    public function show(string $id)
     {
         //
     }
@@ -50,7 +36,7 @@ class PlantController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Plant $plant)
+    public function update(Request $request, string $id)
     {
         //
     }
@@ -58,12 +44,22 @@ class PlantController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Plant $plant)
+    public function destroy(string $id)
     {
+        $plant = Plant::find($id);
+        if (!$plant) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Plant not found'
+                ], 404);
+        }
         if ($plant->image && Storage::disk('public')->exists($plant->image)) {
             Storage::disk('public')->delete($plant->image);
         }
         $plant->delete();
-        return redirect()->route('plants.index')->with('success', 'Plant deleted successfully.');
+        return response()->json([
+            'success' => true,
+            'message' => 'Plant deleted successfully'
+        ]);
     }
 }
